+++ test/interface/golden/__tests__/product/ComplexProduct_spec.ml
@@ -1,5 +1,12 @@
 let () =
   AesonSpec.sampleGoldenAndServerSpec
+    ComplexProduct.decodeSimple
+    ComplexProduct.encodeSimple
+    "simple"
+    "http://localhost:8081/ComplexProduct/Simple"
+    "golden/product/Simple";
+
+  AesonSpec.sampleGoldenAndServerSpec
     ComplexProduct.decodeComplexProduct
     ComplexProduct.encodeComplexProduct
     "complexProduct"