+++ test/interface/golden/product/ComplexProduct.mli
@@ -1,7 +1,19 @@
+type simple =
+  { sa : int
+  ; sb : string
+  }
+
+val encodeSimple : simple -> Js_json.t
+
+val decodeSimple : Js_json.t -> (simple, string) Js_result.t
+
+
 type complexProduct =
-  { cp1 : ((int * (string, float) Aeson.Compatibility.Either.t)) list
+  { cp0 : (Person.person, (int) list) Aeson.Compatibility.Either.t
+  ; cp1 : ((int * (string, float) Aeson.Compatibility.Either.t)) list
   ; cp2 : ((int) list) list
   ; cp3 : ((int) list) option
+  ; cp4 : (simple, int) Aeson.Compatibility.Either.t
   }
 
 val encodeComplexProduct : complexProduct -> Js_json.t